+++ src/app/pages/client/inbox/Notifications.tsx
@@ -480,13 +480,13 @@ function RoomNotificationsGroupComp({
                         src={
                           senderAvatarMxc
                             ? mxcUrlToHttp(
-                                mx,
-                                senderAvatarMxc,
-                                useAuthentication,
-                                48,
-                                48,
-                                'crop'
-                              ) ?? undefined
+                              mx,
+                              senderAvatarMxc,
+                              useAuthentication,
+                              48,
+                              48,
+                              'crop'
+                            ) ?? undefined
                             : undefined
                         }
                         alt={displayName}
@@ -529,9 +529,6 @@ function RoomNotificationsGroupComp({
                     replyEventId={replyEventId}
                     threadRootId={threadRootId}
                     onClick={handleOpenClick}
-                    getMemberPowerTag={getMemberPowerTag}
-                    accessibleTagColors={accessibleTagColors}
-                    legacyUsernameColor={legacyUsernameColor}
                   />
                 )}
                 {renderMatrixEvent(event.type, false, event, displayName, getContent)}