+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.1
 
+- Added name colors and fonts to the member list sidebar.
 - Added a reset button to the room name input box for dms.
   - Reset's the dm room name to the name of the other user (on both ends).
   - Same as saving a blank room name.
@@ -16,3 +17,4 @@
   - Room Font > Space Font > Default
     - _Note, if the room font is invaild it will fallback directly to default, not the space font._
 - Includes all features in Cinny v4.10.3
+- So much refractoring... so much pain...